# [Transform] Fix perf regression in the dynamic case and add a test