d782f54f56def442105656cace5643666acd9b89 introduced a regression while refactoring: the SmallStrategy would be selected by default for dynamic cases.

Fix and add a test.

Ideally we should also have some microbenchmarkings part of presubmit.

Repro:
```
(benchmark-transform-create -r -b cuda  ../iree-samples/transform_dialect/benchmark_linalg_reductions.stub.mlir   reduction_2d_dynamic  f32 128 4096)
```

Before:
```
403.76ms  202.76us              (2 1 1)        (64 1 1)        56        0B        0B  NVIDIA GeForce          1        13                     -                -  reduction_2d_dynamic_dispatch_0_generic_DxD [33]
With transform dialect: reduction_2d_dynamic --function_input="128x4096xf32=1" P50: 202800.00 ns 2.64836291913214990138 GElements/s
```

After:
```
449.43ms  4.8640us            (128 1 1)       (256 1 1)        22        0B  1.0078KB  NVIDIA GeForce          1        13                     -                -  reduction_2d_dynamic_dispatch_0_generic_DxD [33]
With transform dialect: reduction_2d_dynamic --function_input="128x4096xf32=1" P50: 4608.0000 ns 116.55555555555555555555 GElements/s
```