@ftynse 2023-01-20
> Ideally we should also have some microbenchmarkings part of presubmit.

Can't we somehow integrate into the existing benchmarking infra that gets triggered manually when needed?