@silvasean 2023-01-20
I see this more as a "library curation" problem for our set of transform dialect recipes. General benchmarks are unlikely to be the right answer. Probably as we actually start having a "library" of transform dialect recipes we will want to have per-recipe "unit test" benchmarks that do some sort of smoke test of performance in a very directed fashion. The microbenchmarking infra that I'm building could be the basis for this.